+++ src/main/java/com/exactpay/demo/OrderRequest.java
@@ -2,25 +2,8 @@ package com.exactpay.demo;
 
 public class OrderRequest {
 
-    private int totalAmount;
-    private String terminalId;
     private String paymentFlow;
-
-    public int getTotalAmount() {
-        return totalAmount;
-    }
-
-    public void setTotalAmount(int amount) {
-        this.totalAmount = totalAmount;
-    }
-
-    public String getTerminalId() {
-        return terminalId;
-    }
-
-    public void setTerminalId(String terminalId) {
-        this.terminalId = terminalId;
-    }
+    private Details details;
 
     public String getPaymentFlow() {
         return paymentFlow;
@@ -29,9 +12,17 @@ public class OrderRequest {
     public void setPaymentFlow(String paymentFlow) {
         this.paymentFlow = paymentFlow;
     }
-    public OrderRequest(int totalAmount,String terminalId, String paymentFlow) {
-        this.totalAmount = totalAmount;
+    
+    public Details getDetails() {
+        return details;
+    }
+    
+    public void setDetails(Details details) {
+        this.details = details;
+    }
+    
+    public OrderRequest(String paymentFlow, Details details) {
         this.paymentFlow = paymentFlow;
-        this.terminalId = terminalId;
+        this.details = details;
     }
 }
