+++ src/main/resources/application.properties
@@ -1,10 +1,9 @@
-csipay.url = https://staging-pyxisapi.csipay.com/PyxisMasterApi/api/
-security.token=<add Pyxis Security token here>
-server.error.whitelabel.enabled=false
-account.terminalId=<add your terminal ID here>
+csipay.url = https://test-pyxisapi.csipay.com/PyxisMasterApi/api/
+security.token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJDU0lQQVkiLCJpYXQiOjE3NDU5MzkzOTksImV4cCI6MTc0NTk0MDI5OSwidXNlcm5hbWUiOiJhdGhhbyJ9.VPygOFE7VsY50LqWPdGQa1cLcKYDbPO7HM4hzOHjGX0
+account.terminalId=42d13f79eaf3489399fd30f479afd485
 account.paymentFlow=sale
 
 # reCAPTCHA configuration
-google.recaptcha.secret=<add your Google reCaptcha secret key here>
-recaptcha.site-key=<add your Google reCaptcha site key here>
-recaptcha.enabled=true|false
+google.recaptcha.secret=6LcXGw8rAAAAABR-prppv-D7mQcG5JNlY6YJQ2WP
+recaptcha.site-key=6LcXGw8rAAAAAPHaMeglUeXFzkndFr2St6dIBbhL
+recaptcha.enabled=true
