+++ src/main/resources/templates/pay.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>Pay your Order</title>
-    <script src="https://staging-pyxisapi.csipay.com/js/v1/csi_pay.js"></script>
+    <script src="https://test-pyxisapi.csipay.com/js/v1/csi_pay.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL" crossorigin="anonymous"></script>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous">
     <!-- reCAPTCHA script is loaded only if enableRecaptcha is true -->
@@ -125,9 +125,11 @@
         <div class="header-text">
             <h1 class="demo-notice">Demonstration Only</h1>
             <h2 class="demo-subtitle">Payments are simulated and no actual funds are transferred.</h2>
+ <!---
             <h2>
                 <a href="https://developer.exactpay.com/docs/Sandbox-Test-Cards" class="test-cards-link">TEST CARDS</a>
             </h2>
+        -->
             <p class="order-total">Your Order total is: <strong th:text="${amount}"></strong></p>
         </div>
 
@@ -170,7 +172,9 @@
         const components = exact.components({orderId: orderId});
         
         // Setup card element
-        components.addComponent('cardElement', 'full-card', {
+        components.addComponent('cardElement', 'unified', {
+            credit: true,
+            ach: true,
             billingAddress: {
                 type: "full"
             },
@@ -200,10 +204,13 @@
             // Disable button and show processing message
             submitButton.disabled = true;
             processingMessage.style.display = 'block';
+            console.log('---------------- enableRecaptcha ---------')
+            console.log(enableRecaptcha)
             
             // Check if reCAPTCHA is enabled
             if (enableRecaptcha && typeof grecaptcha !== 'undefined') {
                 // Execute reCAPTCHA
+                
                 try {
                     grecaptcha.ready(function() {
                         grecaptcha.execute(recaptchaSiteKey, {action: 'payment_submit'})
@@ -279,6 +286,7 @@
         }
 
         exact.on("payment-complete", (payload) => {
+            console.log('Payment completed ---------------------->');
             document.getElementById('payment_id').value = payload.paymentId;
             window.location = "/confirm?paymentId=" + payload.paymentId;
         });
